requirements.txt
Django==1.11.7
tensorflow==1.4.1
numpy==1.14.0
scipy==1.0.0
scikit_image==0.12.3
opencv_contrib_python==3.4.0.12
matplotlib==2.1.1
Pillow==5.0.0
djangorestframework==3.7.7